refactor: Added shared folder
