+++ src/command/index.js
@@ -2,9 +2,9 @@ import { createInterface } from 'node:readline';
 import { stdin as input, stdout as output } from 'node:process';
 import os from 'os';
 import { operatorsConfig, errors } from '../constants.js';
-import { User } from '../user.js';
+import { User } from '../shared/user.js';
 import { Navigation } from './nwd.js';
-import { Printer } from '../printer.js';
+import { Printer } from '../shared/printer.js';
 import { Hash } from './hash.js';
 import { Zip } from './zip.js';
 import { OperatingSystem } from './os.js';
