+++ src/shared/printer.js
